requirements.txt
tensorflow==2.11.1
keras==2.11.0
numpy==1.24.3
scikit-learn==1.2.2
pandas==2.0.3
matplotlib==3.7.2
scipy==1.10.1
Pillow==10.0.0
vit-keras==0.1.2